Règlage d'un problème j'avais mis les fonction getBoolean et getDirection dans l'interface de la vue de plateau. Maintanant elle sont dans la vue du plateau. Pour facilité leur appel.
