+++ src/Communs/Class/Plateau/PlateauView.java
@@ -1,5 +1,7 @@
 package Communs.Class.Plateau;
 
+import java.util.Scanner;
+
 import Communs.Exceptions.positionInvalide;
 import Communs.Interfaces.Carre;
 import Communs.Interfaces.Plateau.InterfacePlateauView;
@@ -102,5 +104,50 @@ public class PlateauView implements Carre, InterfacePlateauView {
             }
         }
         return res;
+    }
+
+    /**
+     * Methode qui permet de demander une direction dans le terminal.
+     * 
+     * @param sc       System.in permettra de lire la reponse de l'utilisateur
+     * @param question La question qui vas être posee.
+     * @return retourn le cote correpondant a la reponse.
+     */
+    public static Direction demandeDirection(Scanner sc, String question) {
+        System.out.println(question);
+        String demande = sc.nextLine();
+        switch (demande.toUpperCase()) {
+            case "DROITE":
+                return Direction.RIGHT;
+            case "GAUCHE":
+                return Direction.LEFT;
+            case "HAUT":
+                return Direction.UP;
+            case "BAS":
+                return Direction.DOWN;
+            default:
+                return Direction.ACTUEL;
+        }
+    }
+
+    /**
+     * Methode qui permet de demander un boolean.
+     * 
+     * @param sc       System.in permettra de lire la reponse de l'utilisateur
+     * @param question La question qui vas être posee.
+     * @return retourn le boolean correpondant a la reponse.
+     */
+    public static boolean demandeBoolean(Scanner sc, String question) {
+        System.out.println(question);
+        String demande = sc.nextLine();
+        switch (demande.toUpperCase()) {
+            case "OUI":
+                return true;
+            case "NON":
+                return false;
+            default:
+                System.out.println("Erreur : réponse invalide.");
+                return demandeBoolean(sc, question);
+        }
     }
 }
